+++ src/public/javascripts/analytics.js
@@ -1,0 +1,8 @@
+window.dataLayer = window.dataLayer || [];
+
+function gtag() {
+  dataLayer.push(arguments);
+}
+gtag('js', new Date());
+
+gtag('config', 'UA-107716158-1');